refactor: split _fleet.py, _update_checks.py, app.py into focused sub-modules
- _fleet.py (923→341 lines): extract _fleet_display.py (status/watch display),
  _fleet_lifecycle.py (signal guard + reap), _fleet_session.py (session launch)
- _update_checks.py (801→429 lines): extract _update_checks_fetch.py (GitHub HTTP
  cache + fetch machinery) and _update_checks_source.py (source-repo + SHA resolution)
- app.py (740→453 lines): extract _order.py containing order command + 5 private helpers
- Add CD6/CD7 arch guards enforcing new line limits
- Add test_fleet_split.py and test_update_checks_split.py structural tests
