+++ src/autoskillit/cli/_fleet.py
@@ -2,26 +2,26 @@
 
 from __future__ import annotations
 
-import dataclasses
-import fcntl
 import json
 import os
 import shutil
-import signal
 import sys
-import time
-from collections.abc import AsyncIterator
-from contextlib import asynccontextmanager
 from datetime import UTC, datetime
 from pathlib import Path
-from typing import TYPE_CHECKING, Annotated, Literal
+from typing import TYPE_CHECKING, Annotated
 from uuid import uuid4
 
-import anyio
-import anyio.abc
-import psutil
 from cyclopts import App, Parameter
 
+from autoskillit.cli._fleet_display import (
+    _aggregate_totals,
+    _compute_exit_code,
+    _cross_check_tokens,
+    _render_status_display,
+    _watch_loop,
+)
+from autoskillit.cli._fleet_lifecycle import _reap_stale_dispatches
+from autoskillit.cli._fleet_session import _launch_fleet_session
 from autoskillit.core import TerminalColumn, get_logger, is_feature_enabled
 
 logger = get_logger(__name__)
@@ -43,262 +43,7 @@ def _require_fleet(cfg: AutomationConfig) -> None:
 
 if TYPE_CHECKING:
     from autoskillit.config import AutomationConfig
-    from autoskillit.fleet import CampaignState, DispatchRecord, DispatchStatus, ResumeDecision
-    from autoskillit.recipe.schema import Recipe
-
-_STATUS_COLUMNS = (
-    TerminalColumn("NAME", 30, "<"),
-    TerminalColumn("STATUS", 12, "<"),
-    TerminalColumn("ELAPSED", 10, ">"),
-    TerminalColumn("INPUT", 10, ">"),
-    TerminalColumn("OUTPUT", 10, ">"),
-    TerminalColumn("CACHE_RD", 10, ">"),
-    TerminalColumn("CACHE_WR", 10, ">"),
-    TerminalColumn("SESSION_LOG", None, "<"),
-)
-
-
-def _compute_exit_code(state: CampaignState) -> int:
-    """Compute CLI exit code from dispatch statuses.
-
-    0 = all success/skipped, 1 = any failure, 2 = any in-progress.
-    """
-    from autoskillit.fleet import DispatchStatus  # noqa: PLC0415
-
-    _failure = frozenset(
-        {
-            DispatchStatus.FAILURE,
-            DispatchStatus.INTERRUPTED,
-            DispatchStatus.REFUSED,
-            DispatchStatus.RELEASED,
-        }
-    )
-    _in_progress = frozenset(
-        {DispatchStatus.RUNNING, DispatchStatus.PENDING, DispatchStatus.RESUMABLE}
-    )
-    has_failure = any(d.status in _failure for d in state.dispatches)
-    has_in_progress = any(d.status in _in_progress for d in state.dispatches)
-    if has_failure:
-        return 1
-    if has_in_progress:
-        return 2
-    return 0
-
-
-def _fmt_elapsed(dispatch: DispatchRecord) -> str:
-    """Format dispatch elapsed time as human-readable string."""
-    from autoskillit.fleet import DispatchStatus  # noqa: PLC0415
-
-    if dispatch.started_at <= 0:
-        return "-"
-    if dispatch.status == DispatchStatus.RUNNING:
-        seconds = time.time() - dispatch.started_at
-    elif dispatch.ended_at > 0:
-        seconds = dispatch.ended_at - dispatch.started_at
-    else:
-        return "-"
-    seconds = max(0, seconds)
-    if seconds < 60:
-        return f"{seconds:.0f}s"
-    minutes = int(seconds // 60)
-    secs = int(seconds % 60)
-    if minutes < 60:
-        return f"{minutes}m {secs}s"
-    hours = minutes // 60
-    mins = minutes % 60
-    return f"{hours}h {mins}m"
-
-
-def _humanize(n: int | float | None) -> str:
-    """Humanize token counts: 1234 -> '1.2k', 1234567 -> '1.2M'."""
-    if n is None or n == 0:
-        return "0"
-    if not isinstance(n, (int, float)):
-        return "0"
-    if n >= 1_000_000:
-        return f"{n / 1_000_000:.1f}M"
-    if n >= 1_000:
-        return f"{n / 1_000:.1f}k"
-    return str(int(n))
-
-
-def _build_status_rows(state: CampaignState) -> list[tuple[str, ...]]:
-    """Build table rows from campaign state dispatches, including separator and TOTAL rows."""
-    rows: list[tuple[str, ...]] = []
-    for d in state.dispatches:
-        tu = d.token_usage
-        rows.append(
-            (
-                d.name,
-                str(d.status),
-                _fmt_elapsed(d),
-                _humanize(tu.get("input_tokens", 0)),
-                _humanize(tu.get("output_tokens", 0)),
-                _humanize(tu.get("cache_read_input_tokens", 0)),
-                _humanize(tu.get("cache_creation_input_tokens", 0)),
-                d.l2_session_log_dir or "-",
-            )
-        )
-    totals = _aggregate_totals(state)
-    rows.append(("─" * 6, "", "", "", "", "", "", ""))
-    rows.append(
-        (
-            "TOTAL",
-            "",
-            "",
-            _humanize(totals["input_tokens"]),
-            _humanize(totals["output_tokens"]),
-            _humanize(totals["cache_read"]),
-            _humanize(totals["cache_creation"]),
-            "",
-        )
-    )
-    return rows
-
-
-def _aggregate_totals(state: CampaignState) -> dict[str, int]:
-    """Sum token_usage across all dispatches."""
-    totals: dict[str, int] = {
-        "input_tokens": 0,
-        "output_tokens": 0,
-        "cache_read": 0,
-        "cache_creation": 0,
-    }
-    for d in state.dispatches:
-        tu = d.token_usage
-        totals["input_tokens"] += tu.get("input_tokens", 0)
-        totals["output_tokens"] += tu.get("output_tokens", 0)
-        totals["cache_read"] += tu.get("cache_read_input_tokens", 0)
-        totals["cache_creation"] += tu.get("cache_creation_input_tokens", 0)
-    return totals
-
-
-def _load_log_totals(state: CampaignState) -> dict[str, int] | None:
-    """Load token totals from sessions.jsonl for the campaign. Returns None if unavailable."""
-    from autoskillit.execution import resolve_log_dir
-    from autoskillit.pipeline import DefaultTokenLog
-
-    log_root = resolve_log_dir("")
-    sessions_index = log_root / "sessions.jsonl"
-    if not sessions_index.exists():
-        return None
-
-    token_log = DefaultTokenLog()
-    loaded = token_log.load_from_log_dir(log_root, campaign_id_filter=state.campaign_id)
-    if loaded == 0:
-        return None
-
-    return token_log.compute_total()
-
-
-def _cross_check_tokens(state: CampaignState, state_totals: dict[str, int]) -> None:
-    """Warn on >5% token divergence between state.json and sessions.jsonl."""
-    log_totals = _load_log_totals(state)
-    if log_totals is None:
-        return
-
-    for label, state_key, log_key in [
-        ("input_tokens", "input_tokens", "input_tokens"),
-        ("output_tokens", "output_tokens", "output_tokens"),
-        ("cache_read", "cache_read", "cache_read_input_tokens"),
-        ("cache_creation", "cache_creation", "cache_creation_input_tokens"),
-    ]:
-        sv = state_totals.get(state_key, 0)
-        lv = log_totals.get(log_key, 0)
-        if sv > 0 and abs(sv - lv) / sv > 0.05:
-            sys.stderr.write(
-                f"WARNING: {label} diverge {abs(sv - lv) / sv:.1%} (>5%)"
-                f" (state={sv}, sessionlog={lv}); state.json wins\n"
-            )
-
-
-def _render_status_display(state: CampaignState) -> int:
-    """Print campaign header and 8-column dispatch table to stdout.
-
-    Returns the number of lines printed (for cursor-based screen refresh).
-    """
-    from autoskillit.cli._ansi import _render_terminal_table
-
-    started = datetime.fromtimestamp(state.started_at, tz=UTC).strftime("%Y-%m-%d %H:%M:%S UTC")
-    header = f"Campaign: {state.campaign_name}  ID: {state.campaign_id}  Started: {started}"
-    print(header)
-
-    rows = _build_status_rows(state)
-    table_str = _render_terminal_table(_STATUS_COLUMNS, rows)
-    print(table_str)
-    return 1 + table_str.rstrip("\n").count("\n") + 1
-
-
-def _watch_loop(state_path: Path) -> int:
-    """1 Hz polling loop for fleet status. Returns exit code."""
-    import select
-    import termios
-    import tty
-
-    from autoskillit.fleet import DispatchStatus, read_state  # noqa: PLC0415
-
-    _in_progress = frozenset({DispatchStatus.RUNNING, DispatchStatus.PENDING})
-
-    state = read_state(state_path)
-    if state is None:
-        sys.stderr.write("ERROR: state file disappeared or corrupted\n")
-        return 3
-
-    # If campaign already terminal, render once and exit without needing a TTY
-    if all(d.status not in _in_progress for d in state.dispatches):
-        _render_status_display(state)
-        print("\nAll dispatches complete.")
-        return _compute_exit_code(state)
-
-    if not sys.stdin.isatty() or not sys.stdout.isatty():
-        sys.stderr.write(
-            "ERROR: --watch requires an interactive terminal"
-            " (both stdin and stdout must be TTYs).\n"
-        )
-        return 1
-
-    fd = sys.stdin.fileno()
-    old_settings = termios.tcgetattr(fd)
-    prev_lines = 0
-    try:
-        tty.setcbreak(fd)
-        while True:
-            state = read_state(state_path)
-            if state is None:
-                sys.stdout.flush()
-                sys.stderr.write("ERROR: state file disappeared or corrupted\n")
-                return 3
-
-            if prev_lines > 0:
-                sys.stdout.write(f"\033[{prev_lines}A")
-                for _ in range(prev_lines):
-                    sys.stdout.write("\033[2K\033[1B")
-                sys.stdout.write(f"\033[{prev_lines}A")
-            sys.stdout.flush()
-
-            prev_lines = _render_status_display(state)
-
-            if all(d.status not in _in_progress for d in state.dispatches):
-                print("\nAll dispatches complete.")
-                return _compute_exit_code(state)
-
-            rlist, _, _ = select.select([sys.stdin], [], [], 1.0)
-            if rlist:
-                ch = sys.stdin.read(1)
-                if ch.lower() == "q":
-                    return _compute_exit_code(state)
-    except KeyboardInterrupt:
-        state = read_state(state_path)
-        return _compute_exit_code(state) if state else 3
-    except Exception as exc:
-        logger.error("unexpected error in --watch loop: %s", exc, exc_info=True)
-        sys.stderr.write(f"ERROR: unexpected error in --watch loop: {exc}\n")
-        return 3
-    finally:
-        try:
-            termios.tcsetattr(fd, termios.TCSADRAIN, old_settings)
-        except (termios.error, OSError):
-            pass
+    from autoskillit.fleet import ResumeDecision
 
 
 fleet_app = App(name="fleet", help="Campaign fleet management.")
@@ -314,326 +59,6 @@ def _remove_clone_fn(path: str, _flag: str) -> dict[str, str]:
         return {"removed": "false", "reason": str(exc)}
 
 
-def _launch_fleet_session(
-    campaign_recipe: Recipe | None,
-    campaign_id: str | None,
-    state_path: Path | None,
-    resume_metadata: ResumeDecision | None,
-    *,
-    fleet_mode: Literal["dispatch", "campaign"],
-) -> None:
-    """Build the L3 orchestrator prompt and launch an interactive fleet session."""
-    from autoskillit.cli import detect_autoskillit_mcp_prefix  # noqa: PLC0415
-    from autoskillit.cli._session_launch import _run_interactive_session
-
-    mcp_prefix = detect_autoskillit_mcp_prefix()
-
-    from autoskillit.core import NoResume
-
-    project_dir = Path.cwd()
-
-    if campaign_recipe is None:
-        # Ad-hoc mode: no campaign, no state, bare kitchen open
-        from autoskillit.cli._prompts import _build_fleet_dispatch_prompt
-
-        prompt = _build_fleet_dispatch_prompt(mcp_prefix)
-        extra_env: dict[str, str] = {
-            "AUTOSKILLIT_SESSION_TYPE": "fleet",
-            "AUTOSKILLIT_FLEET_MODE": fleet_mode,
-            "AUTOSKILLIT_HEADLESS": "0",
-        }
-        while True:
-            reload_id = _run_interactive_session(
-                prompt, extra_env=extra_env, resume_spec=NoResume(), project_dir=project_dir
-            )
-            if reload_id is None:
-                break
-    else:
-        # Campaign-driven mode: full orchestrator prompt with manifest and state
-        if campaign_id is None:
-            raise ValueError("campaign_id must not be None in campaign-driven mode")
-        if state_path is None:
-            raise ValueError("state_path must not be None in campaign-driven mode")
-        from autoskillit.cli._prompts import _build_fleet_campaign_prompt
-        from autoskillit.core import dump_yaml_str
-
-        manifest_yaml = dump_yaml_str(
-            [dataclasses.asdict(d) for d in campaign_recipe.dispatches],
-            default_flow_style=False,
-            allow_unicode=True,
-        )
-        completed_dispatches = (
-            resume_metadata.completed_dispatches_block if resume_metadata is not None else ""
-        )
-        resumable_dispatch_name = (
-            resume_metadata.next_dispatch_name
-            if resume_metadata is not None and resume_metadata.is_resumable
-            else ""
-        )
-        prompt = _build_fleet_campaign_prompt(
-            campaign_recipe,
-            manifest_yaml,
-            completed_dispatches,
-            mcp_prefix,
-            campaign_id,
-            resumable_dispatch_name=resumable_dispatch_name,
-        )
-        extra_env = {
-            "AUTOSKILLIT_SESSION_TYPE": "fleet",
-            "AUTOSKILLIT_FLEET_MODE": fleet_mode,
-            "AUTOSKILLIT_CAMPAIGN_ID": campaign_id,
-            "AUTOSKILLIT_CAMPAIGN_STATE_PATH": str(state_path),
-            "AUTOSKILLIT_HEADLESS": "0",
-        }
-        while True:
-            reload_id = _run_interactive_session(
-                prompt, extra_env=extra_env, resume_spec=NoResume(), project_dir=project_dir
-            )
-            if reload_id is None:
-                break
-
-
-def _transition_dead_dispatch(
-    state_path: Path, record: DispatchRecord, reason: str
-) -> DispatchStatus | None:
-    """Transition a confirmed-dead RUNNING dispatch to RESUMABLE or INTERRUPTED.
-
-    Delegates sidecar decision logic to the fleet layer. Returns the new status,
-    or None if both write attempts failed. Never raises.
-    """
-    from autoskillit.fleet import crash_recover_dispatch  # noqa: PLC0415
-
-    return crash_recover_dispatch(state_path, record, reason=reason)
-
-
-@asynccontextmanager
-async def _fleet_signal_guard(
-    state_path: Path,
-    campaign_id: str,
-    *,
-    campaign_name: str | None = None,
-    cleanup_on_interrupt: bool = False,
-) -> AsyncIterator[None]:
-    """Async context manager that installs SIGINT/SIGTERM handlers.
-
-    On signal receipt:
-    - Cancels the enclosing task group scope first.
-    - Reads state.json and marks RUNNING dispatches as INTERRUPTED.
-    - Verifies PID identity via starttime_ticks before killing.
-    - Optionally runs workspace cleanup.
-    - Logs a resume hint.
-    """
-
-    async def _watch(
-        scope: anyio.CancelScope,
-        *,
-        task_status: anyio.abc.TaskStatus = anyio.TASK_STATUS_IGNORED,
-    ) -> None:
-        with anyio.open_signal_receiver(signal.SIGTERM, signal.SIGINT, signal.SIGHUP) as signals:
-            task_status.started()
-            async for sig in signals:
-                signame = sig.name
-
-                # Cancel the enclosing scope FIRST to unwind any in-flight dispatch
-                # coroutine before the cleanup writes state (prevents ordering races).
-                scope.cancel()
-
-                # Shield the cleanup from the now-cancelled scope so that async
-                # operations (kill, state write) are not interrupted.
-                with anyio.CancelScope(shield=True):
-                    from autoskillit.execution import async_kill_process_tree, read_starttime_ticks
-                    from autoskillit.fleet import (  # noqa: PLC0415
-                        DispatchStatus,
-                        read_state,
-                    )
-
-                    state = read_state(state_path)
-                    if state is not None:
-                        for dispatch in state.dispatches:
-                            if dispatch.status != DispatchStatus.RUNNING:
-                                continue
-                            if dispatch.l2_pid == 0:
-                                _transition_dead_dispatch(
-                                    state_path, dispatch, reason=f"signal_{signame}"
-                                )
-                                continue
-
-                            # Verify PID identity before killing
-                            current_ticks = read_starttime_ticks(dispatch.l2_pid)
-                            if current_ticks is not None:
-                                if (
-                                    dispatch.l2_starttime_ticks > 0
-                                    and current_ticks == dispatch.l2_starttime_ticks
-                                ):
-                                    try:
-                                        await async_kill_process_tree(dispatch.l2_pid, timeout=5.0)
-                                    except Exception:
-                                        logger.warning(
-                                            "signal_guard: kill_process_tree failed",
-                                            exc_info=True,
-                                        )
-                                else:
-                                    logger.warning(
-                                        "signal_guard: PID %d recycled (ticks mismatch)",
-                                        dispatch.l2_pid,
-                                    )
-                            else:
-                                # Non-Linux fallback: psutil.pid_exists without identity check
-                                if psutil.pid_exists(dispatch.l2_pid):
-                                    try:
-                                        await async_kill_process_tree(dispatch.l2_pid, timeout=5.0)
-                                    except Exception:
-                                        logger.warning(
-                                            "signal_guard: kill_process_tree failed (non-linux)",
-                                            exc_info=True,
-                                        )
-
-                            _transition_dead_dispatch(
-                                state_path, dispatch, reason=f"signal_{signame}"
-                            )
-
-                    if cleanup_on_interrupt:
-                        try:
-                            from autoskillit.core import ensure_project_temp
-                            from autoskillit.workspace import DefaultWorkspaceManager
-
-                            workspace_dir = ensure_project_temp(Path.cwd())
-                            mgr = DefaultWorkspaceManager()
-                            mgr.delete_contents(workspace_dir)
-                        except Exception:
-                            logger.warning("signal_guard: workspace cleanup failed", exc_info=True)
-
-                    if campaign_name is not None:
-                        resume_cmd = (
-                            f"autoskillit fleet campaign {campaign_name} --resume {campaign_id}"
-                        )
-                    else:
-                        resume_cmd = f"autoskillit fleet campaign <name> --resume {campaign_id}"
-                    sys.stderr.write(f"Campaign {campaign_id} interrupted. Resume: {resume_cmd}\n")
-                return
-
-    async with anyio.create_task_group() as tg:
-        await tg.start(_watch, tg.cancel_scope)
-        try:
-            yield
-        finally:
-            tg.cancel_scope.cancel()
-
-
-def _reap_stale_dispatches(state_path: Path, *, dry_run: bool = False) -> None:
-    """Reap stale RUNNING dispatches with PID-recycling-safe identity checks.
-
-    Uses fcntl.flock() to protect against concurrent reap invocations.
-    For each RUNNING dispatch:
-    - Boot-ID mismatch → reaped_pid_recycled (no kill)
-    - Process dead → reaped_dead_pid
-    - Process alive + ticks match → kill + reaped_orphan
-    - Process alive + ticks mismatch → reaped_pid_recycled (no kill)
-    """
-    from autoskillit.execution import kill_process_tree, read_boot_id, read_starttime_ticks
-    from autoskillit.fleet import DispatchStatus, read_state  # noqa: PLC0415
-
-    current_boot_id = read_boot_id()
-
-    if not state_path.exists():
-        logger.info("reap: state file not found, nothing to reap: %s", state_path)
-        return
-
-    with open(state_path, "r+") as _lock_fh:
-        fcntl.flock(_lock_fh, fcntl.LOCK_EX)
-        try:
-            state = read_state(state_path)
-            if state is None:
-                logger.warning("reap: cannot read state file: %s", state_path)
-                return
-
-            running = [d for d in state.dispatches if d.status == DispatchStatus.RUNNING]
-            if not running:
-                logger.info("reap: no running dispatches in campaign %s", state.campaign_id)
-                return
-
-            logger.info(
-                "reap: scanning %d dispatches in campaign %s", len(running), state.campaign_id
-            )
-
-            for dispatch in running:
-                name = dispatch.name
-                pid = dispatch.l2_pid
-
-                if pid == 0:
-                    if dry_run:
-                        logger.info("reap: [WOULD MARK]  %s  pid=0  (no PID recorded)", name)
-                    else:
-                        new_status = _transition_dead_dispatch(
-                            state_path, dispatch, reason="reaped_dead_pid"
-                        )
-                        if new_status is not None:
-                            logger.info("reap: [MARKED]      %s  (no PID recorded)", name)
-                        else:
-                            logger.warning("reap: [MARK_FAILED] %s  (no PID recorded)", name)
-                    continue
-
-                # Boot ID check: if machine rebooted, all PIDs are recycled
-                if (
-                    dispatch.l2_boot_id != ""
-                    and current_boot_id is not None
-                    and dispatch.l2_boot_id != current_boot_id
-                ):
-                    if dry_run:
-                        logger.info(
-                            "reap: [WOULD MARK]  %s  pid=%d  (rebooted, pid_recycled)", name, pid
-                        )
-                    else:
-                        _transition_dead_dispatch(
-                            state_path, dispatch, reason="reaped_pid_recycled"
-                        )
-                        logger.info(
-                            "reap: [MARKED]      %s  pid=%d  (rebooted, pid_recycled)",
-                            name,
-                            pid,
-                        )
-                    continue
-
-                if not psutil.pid_exists(pid):
-                    if dry_run:
-                        logger.info("reap: [WOULD MARK]  %s  pid=%d  (process dead)", name, pid)
-                    else:
-                        _transition_dead_dispatch(state_path, dispatch, reason="reaped_dead_pid")
-                        logger.info("reap: [MARKED]      %s  pid=%d  (process dead)", name, pid)
-                    continue
-
-                # Process is alive — check identity
-                current_ticks = read_starttime_ticks(pid)
-                if current_ticks is not None and current_ticks == dispatch.l2_starttime_ticks:
-                    if dry_run:
-                        logger.info(
-                            "reap: [WOULD KILL]  %s  pid=%d  (orphan, identity match)", name, pid
-                        )
-                    else:
-                        try:
-                            kill_process_tree(pid)
-                        except Exception:
-                            logger.warning(
-                                "reap: kill_process_tree failed for pid=%d", pid, exc_info=True
-                            )
-                        _transition_dead_dispatch(state_path, dispatch, reason="reaped_orphan")
-                        logger.info("reap: [KILLED]      %s  pid=%d  (orphan reaped)", name, pid)
-                else:
-                    if dry_run:
-                        logger.info(
-                            "reap: [WOULD MARK]  %s  pid=%d  (PID recycled, no kill)", name, pid
-                        )
-                    else:
-                        _transition_dead_dispatch(
-                            state_path, dispatch, reason="reaped_pid_recycled"
-                        )
-                        logger.info(
-                            "reap: [MARKED]      %s  pid=%d  (PID recycled, no kill)",
-                            name,
-                            pid,
-                        )
-        finally:
-            fcntl.flock(_lock_fh, fcntl.LOCK_UN)
 
 
 @fleet_app.command(name="dispatch")
