+++ src/autoskillit/cli/_update_checks.py
@@ -4,13 +4,11 @@ Consolidates version-staleness, hook-drift, and source-drift signals into a
 single dismissable prompt per CLI invocation.
 
 Branch-aware dismissal windows:
+- stable/main/release-tag/UNKNOWN: timedelta(days=7)
+- integration/LOCAL_EDITABLE: timedelta(hours=12)
 
-- ``stable`` / ``main`` / release-tag / ``UNKNOWN`` installs: ``timedelta(days=7)``
-- ``integration`` / ``LOCAL_EDITABLE`` installs: ``timedelta(hours=12)``
-
-Dismissal expires on two axes (see ``_is_dismissed``):
-1. Time window elapsed (branch-aware, not SHA-keyed).
-2. Version advanced past the ``dismissed_version`` recorded at dismiss time.
+Dismissal expires on two axes: time window elapsed, or version advanced past
+the dismissed_version recorded at dismiss time.
 """
 
 from __future__ import annotations
@@ -19,13 +17,10 @@ import json
 import os
 import subprocess
 import sys
-import time
 from dataclasses import dataclass
 from datetime import UTC, datetime, timedelta
 from pathlib import Path
-from typing import Any, Literal
-
-import httpx
+from typing import Literal
 
 from autoskillit.cli._hooks import _claude_settings_path
 from autoskillit.cli._install_info import (
@@ -38,7 +33,13 @@ from autoskillit.cli._install_info import (
 )
 from autoskillit.cli._restart import perform_restart
 from autoskillit.cli._terminal import terminal_guard
-from autoskillit.core import AUTOSKILLIT_INSTALLED_VERSION, atomic_write, get_logger
+from autoskillit.cli._update_checks_fetch import (
+    _fetch_latest_version,
+    _fetch_with_cache,  # noqa: F401 re-exported for test backward compat
+    invalidate_fetch_cache,
+)
+from autoskillit.cli._update_checks_source import resolve_reference_sha
+from autoskillit.core import atomic_write, get_logger
 from autoskillit.hook_registry import _count_hook_registry_drift
 
 logger = get_logger(__name__)
@@ -47,27 +48,6 @@ _DISMISS_FILE = "update_check.json"
 _STABLE_DISMISS_WINDOW = timedelta(days=7)
 _DEV_DISMISS_WINDOW = timedelta(hours=12)
 
-# Disposable on-disk cache for GitHub API GETs.  This is intentionally a plain
-# JSON dict (not ``write_versioned_json``) — it is a transient performance
-# cache, not a persisted schema artifact, so a stray pre-existing file at the
-# same path is safely ignored on a JSON parse failure.
-_FETCH_CACHE_FILE = "github_fetch_cache.json"
-_DEFAULT_FETCH_TTL_SECONDS = 30 * 60
-
-# connect=2s buffers for cold DNS (httpx's connect timeout does NOT cover the
-# OS resolver itself); read=1s is tight because 304 responses are tiny.  The
-# 30-min disk TTL above is the **only reliable quota defense** — GitHub's
-# x-ratelimit-used counter has been observed to increment on 304 responses
-# despite their docs claiming otherwise, so ETag/If-None-Match is a
-# bandwidth/latency optimization, not a rate-limit optimization.
-_HTTP_TIMEOUT = httpx.Timeout(connect=2.0, read=1.0, write=5.0, pool=1.0)
-_GITHUB_API_VERSION = "2022-11-28"
-
-_GITHUB_RELEASES_URL = "https://api.github.com/repos/TalonT-Org/AutoSkillit/releases/latest"
-_GITHUB_INTEGRATION_PYPROJECT_URL = (
-    "https://api.github.com/repos/TalonT-Org/AutoSkillit/contents/pyproject.toml?ref=integration"
-)
-
 
 @dataclass(frozen=True)
 class Signal:
@@ -75,11 +55,6 @@ class Signal:
 
     kind: Literal["binary", "hooks", "source_drift", "dual_mcp"]
     message: str
-
-
-# ---------------------------------------------------------------------------
-# Dismiss state I/O (ported from _stale_check.py)
-# ---------------------------------------------------------------------------
 
 
 def _read_dismiss_state(home: Path) -> dict[str, object]:
@@ -101,333 +76,6 @@ def _write_dismiss_state(home: Path, state: dict[str, object]) -> None:
     state_file = home / ".autoskillit" / _DISMISS_FILE
     state_file.parent.mkdir(parents=True, exist_ok=True)
     atomic_write(state_file, json.dumps(state))
-
-
-# ---------------------------------------------------------------------------
-# Fetch cache (ported from _stale_check.py)
-# ---------------------------------------------------------------------------
-
-
-def _fetch_cache_path(home: Path) -> Path:
-    return home / ".autoskillit" / _FETCH_CACHE_FILE
-
-
-def _read_fetch_cache(home: Path) -> dict[str, Any]:
-    """Read the GitHub fetch cache; tolerate any read/parse failure."""
-    try:
-        data = json.loads(_fetch_cache_path(home).read_text(encoding="utf-8"))
-        return data if isinstance(data, dict) else {}
-    except Exception:
-        logger.debug("Failed to read fetch cache", exc_info=True)
-        return {}
-
-
-def _write_fetch_cache(home: Path, data: dict[str, Any]) -> None:
-    """Write the GitHub fetch cache atomically."""
-    target = _fetch_cache_path(home)
-    target.parent.mkdir(parents=True, exist_ok=True)
-    try:
-        atomic_write(target, json.dumps(data))
-    except Exception:
-        logger.debug("Failed to write fetch cache", exc_info=False)
-
-
-def invalidate_fetch_cache(home: Path) -> None:
-    """Delete the GitHub fetch cache. Call after install/update."""
-    try:
-        _fetch_cache_path(home).unlink(missing_ok=True)
-    except OSError:
-        logger.debug("Failed to invalidate fetch cache", exc_info=True)
-
-
-def _scrub_auth(text: str) -> str:
-    """Remove any GITHUB_TOKEN value or ``Bearer …`` token from a string."""
-    if not text:
-        return text
-    token = os.environ.get("GITHUB_TOKEN")
-    if token:
-        text = text.replace(token, "***")
-    import re
-
-    text = re.sub(r"(?i)Bearer\s+\S+", "Bearer ***", text)
-    return text
-
-
-def _resolve_fetch_ttl() -> int:
-    raw = os.environ.get("AUTOSKILLIT_FETCH_CACHE_TTL_SECONDS")
-    if raw is None:
-        return _DEFAULT_FETCH_TTL_SECONDS
-    try:
-        return int(raw)
-    except ValueError:
-        return _DEFAULT_FETCH_TTL_SECONDS
-
-
-def _fetch_with_cache(url: str, *, home: Path, ttl: int | None = None) -> dict[str, Any] | None:
-    """Fetch ``url`` via httpx with a disk-backed cache and conditional GETs.
-
-    Returns the parsed JSON dict on success (200 or 304), or ``None`` on any
-    error.  ``Authorization`` headers are scrubbed from log output to prevent
-    token leakage via DEBUG-level error reporting.
-
-    The 30-minute disk TTL is the only reliable quota defense — see
-    ``_DEFAULT_FETCH_TTL_SECONDS`` for the rationale.
-    """
-    effective_ttl = ttl if ttl is not None else _resolve_fetch_ttl()
-    cache = _read_fetch_cache(home)
-    entry = cache.get(url) if isinstance(cache.get(url), dict) else None
-    now = time.time()
-    if entry is not None:
-        cached_at = entry.get("cached_at")
-        body = entry.get("body")
-        if isinstance(cached_at, (int, float)) and isinstance(body, dict):
-            if now - cached_at < effective_ttl:
-                if entry.get("installed_version") == AUTOSKILLIT_INSTALLED_VERSION:
-                    return body
-
-    headers: dict[str, str] = {
-        "Accept": "application/vnd.github+json",
-        "X-GitHub-Api-Version": _GITHUB_API_VERSION,
-        "User-Agent": f"autoskillit/{AUTOSKILLIT_INSTALLED_VERSION}",
-    }
-    token = os.environ.get("GITHUB_TOKEN")
-    if token:
-        headers["Authorization"] = f"Bearer {token}"
-    if entry is not None and isinstance(entry.get("etag"), str):
-        headers["If-None-Match"] = entry["etag"]
-
-    try:
-        with httpx.Client(timeout=_HTTP_TIMEOUT) as client:
-            response = client.get(url, headers=headers)
-    except Exception as exc:
-        logger.debug("fetch failed for %s: %s", url, _scrub_auth(str(exc)))
-        return None
-
-    try:
-        if response.status_code == 304 and entry is not None:
-            body = entry.get("body")
-            if isinstance(body, dict):
-                cache[url] = {
-                    "body": body,
-                    "etag": entry.get("etag"),
-                    "cached_at": now,
-                    "installed_version": AUTOSKILLIT_INSTALLED_VERSION,
-                }
-                _write_fetch_cache(home, cache)
-                return body
-            return None
-        if response.status_code == 200:
-            body = response.json()
-            if not isinstance(body, dict):
-                return None
-            etag = response.headers.get("ETag")
-            cache[url] = {
-                "body": body,
-                "etag": etag,
-                "cached_at": now,
-                "installed_version": AUTOSKILLIT_INSTALLED_VERSION,
-            }
-            _write_fetch_cache(home, cache)
-            return body
-        logger.debug("fetch %s returned status %d", url, response.status_code)
-        return None
-    except Exception as exc:
-        logger.debug("fetch parse failed for %s: %s", url, _scrub_auth(str(exc)))
-        return None
-
-
-def _fetch_latest_version(target: str, home: Path) -> str | None:
-    """Fetch the latest available version for the given target branch.
-
-    ``target`` is either ``"releases/latest"`` (for stable/main installs) or
-    ``"integration"`` (for integration installs).
-
-    Returns ``None`` on any network error or timeout.
-    """
-    try:
-        if target == "integration":
-            data = _fetch_with_cache(_GITHUB_INTEGRATION_PYPROJECT_URL, home=home)
-            if data is None:
-                return None
-            import base64
-
-            raw_content = data.get("content")
-            if not isinstance(raw_content, str):
-                return None
-            content = base64.b64decode(raw_content).decode("utf-8")
-            for line in content.splitlines():
-                line = line.strip()
-                if line.split("=", 1)[0].strip() == "version":
-                    return line.split("=", 1)[1].strip().strip('"').strip("'")
-            return None
-        data = _fetch_with_cache(_GITHUB_RELEASES_URL, home=home)
-        if data is None:
-            return None
-        tag = data.get("tag_name", "")
-        return tag.lstrip("v") if isinstance(tag, str) and tag else None
-    except Exception as exc:
-        logger.debug("Failed to fetch latest version from GitHub: %s", _scrub_auth(str(exc)))
-        return None
-
-
-# ---------------------------------------------------------------------------
-# Source-repo discovery and SHA resolution (ported from _source_drift.py)
-# ---------------------------------------------------------------------------
-
-
-def find_source_repo() -> Path | None:
-    """Locate the autoskillit source repository root.
-
-    Resolution order:
-    1. ``AUTOSKILLIT_SOURCE_REPO`` env var (must exist and contain ``src/autoskillit/``).
-    2. Walk upward from ``Path.cwd()``, match ``pyproject.toml``
-       ``[project].name == "autoskillit"`` AND ``src/autoskillit/`` present.
-
-    Returns ``None`` if no match found or on any error.
-    """
-    try:
-        env_val = os.environ.get("AUTOSKILLIT_SOURCE_REPO")
-        if env_val:
-            candidate = Path(env_val)
-            if candidate.exists() and (candidate / "src" / "autoskillit").exists():
-                return candidate
-            logger.debug(
-                "AUTOSKILLIT_SOURCE_REPO=%s not usable (missing or no src/autoskillit/), "
-                "falling through to CWD walk",
-                env_val,
-            )
-
-        import tomllib
-
-        current = Path.cwd()
-        while True:
-            pyproject = current / "pyproject.toml"
-            if pyproject.is_file():
-                try:
-                    with open(pyproject, "rb") as fh:
-                        data = tomllib.load(fh)
-                    project_name = data.get("project", {}).get("name")
-                    if (
-                        project_name == "autoskillit"
-                        and (current / "src" / "autoskillit").exists()
-                    ):
-                        return current
-                except Exception:
-                    logger.debug("drift check: could not parse %s", pyproject, exc_info=True)
-
-            parent = current.parent
-            if parent == current:  # Filesystem root
-                break
-            current = parent
-
-        return None
-    except Exception:
-        logger.debug("drift check: find_source_repo failed", exc_info=True)
-        return None
-
-
-def resolve_reference_sha(
-    info: InstallInfo,
-    home: Path,
-    *,
-    network: bool = True,
-) -> str | None:
-    """Resolve the current HEAD SHA of the branch the install was tracking.
-
-    Returns ``None`` when the SHA cannot be determined (network offline, no
-    source repo, unknown revision).  The caller treats ``None`` as "skip check"
-    (fail-open).
-
-    Args:
-        info: Install classification from ``detect_install()``.
-        home: User home directory (used by the disk-backed fetch cache).
-        network: When ``False``, only the local git or disk cache is consulted.
-            The doctor check passes ``network=True`` so it can resolve remote refs.
-    """
-    try:
-        if info.requested_revision is None:
-            logger.debug("drift check skipped: no requested_revision in direct_url.json")
-            return None
-
-        rev = info.requested_revision
-
-        # Short-circuit: exact SHA equality means no drift is possible.
-        # IMPORTANT: use == not startswith — a branch named after a hex prefix
-        # of the commit SHA must NOT false-positive here.
-        if rev == info.commit_id:
-            return info.commit_id
-
-        sha: str | None = None
-
-        source_repo = find_source_repo()
-        if source_repo is not None and source_repo.exists():
-            sha = _git_ls_remote_sha(source_repo, rev)
-
-        if sha is None:
-            sha = _api_sha(rev, home, network=network)
-
-        return sha
-
-    except Exception:
-        logger.debug("drift check skipped: resolve_reference_sha error", exc_info=True)
-        return None
-
-
-def _git_ls_remote_sha(source_repo: Path, rev: str) -> str | None:
-    """Run git ls-remote to resolve a branch or tag ref SHA.
-
-    Tries ``refs/heads/<rev>`` first, then ``refs/tags/<rev>^{}`` for peeled
-    tag objects.  Returns ``None`` on empty output or any subprocess error.
-    """
-    for ref in (f"refs/heads/{rev}", f"refs/tags/{rev}^{{}}"):
-        try:
-            result = subprocess.run(
-                ["git", "-C", str(source_repo), "ls-remote", "origin", ref],
-                capture_output=True,
-                text=True,
-                timeout=5,
-                env=os.environ,
-            )
-            first_line = result.stdout.strip().splitlines()[0] if result.stdout.strip() else ""
-            if first_line:
-                sha = first_line.split()[0]
-                if sha:
-                    return sha
-        except (subprocess.SubprocessError, FileNotFoundError, OSError, IndexError):
-            logger.debug("git ls-remote failed for ref=%s", ref, exc_info=True)
-    return None
-
-
-def _api_sha(rev: str, home: Path, *, network: bool = True) -> str | None:
-    """Return the commit SHA for ``rev`` from the GitHub API or disk cache.
-
-    When ``network=False`` (doctor mode), reads the existing cache without
-    any TTL check and makes no outbound HTTP request.  Returns ``None`` if the
-    cache has no entry for the URL.
-    """
-    # Try refs/heads first; fall back to refs/tags for tag revisions.
-    ref_prefix = "refs/tags" if rev.startswith("v") else "refs/heads"
-    url = f"https://api.github.com/repos/TalonT-Org/AutoSkillit/git/{ref_prefix}/{rev}"
-
-    if network:
-        data: Any = _fetch_with_cache(url, home=home)
-    else:
-        cache = _read_fetch_cache(home)
-        entry = cache.get(url) if isinstance(cache.get(url), dict) else None
-        data = entry.get("body") if isinstance(entry, dict) else None
-
-    if not isinstance(data, dict):
-        return None
-    obj = data.get("object")
-    if not isinstance(obj, dict):
-        return None
-    sha = obj.get("sha")
-    return sha if isinstance(sha, str) and sha else None
-
-
-# ---------------------------------------------------------------------------
-# Update result verification (ported from _stale_check.py)
-# ---------------------------------------------------------------------------
 
 
 def _verify_update_result(
@@ -470,11 +118,6 @@ def _verify_update_result(
         flush=True,
     )
     return False
-
-
-# ---------------------------------------------------------------------------
-# Signal gatherers
-# ---------------------------------------------------------------------------
 
 
 def _binary_signal(info: InstallInfo, home: Path, current: str) -> Signal | None:
@@ -561,11 +204,6 @@ def _dual_mcp_signal(home: Path | None = None) -> Signal | None:
     return None
 
 
-# ---------------------------------------------------------------------------
-# Unified dismissal
-# ---------------------------------------------------------------------------
-
-
 def _is_dismissed(
     state: dict[str, object],
     *,
@@ -613,11 +251,6 @@ def _is_dismissed(
         return False
 
 
-# ---------------------------------------------------------------------------
-# Update sequence
-# ---------------------------------------------------------------------------
-
-
 def _run_update_sequence(
     info: InstallInfo,
     current: str,
@@ -655,11 +288,6 @@ def _run_update_sequence(
         _write_dismiss_state(home, state)
         invalidate_fetch_cache(home)
         perform_restart()
-
-
-# ---------------------------------------------------------------------------
-# Main entry point
-# ---------------------------------------------------------------------------
 
 
 def run_update_checks(home: Path | None = None) -> None:
