+++ tests/arch/test_cli_decomposition.py
@@ -87,6 +87,22 @@ def test_doctor_py_under_line_limit():
     assert len(lines) <= 250, f"_doctor.py is {len(lines)} lines — split required"
 
 
+# CD6
+def test_fleet_py_under_line_limit():
+    """CD6: _fleet.py must be ≤400 lines after sub-module extraction."""
+    p = Path(__file__).parent.parent.parent / "src" / "autoskillit" / "cli" / "_fleet.py"
+    lines = len(p.read_text().splitlines())
+    assert lines <= 400, f"_fleet.py is {lines} lines — extract display/lifecycle/session"
+
+
+# CD7
+def test_update_checks_py_under_line_limit():
+    """CD7: _update_checks.py must be ≤450 lines after sub-module extraction."""
+    p = Path(__file__).parent.parent.parent / "src" / "autoskillit" / "cli" / "_update_checks.py"
+    lines = len(p.read_text().splitlines())
+    assert lines <= 450, f"_update_checks.py is {lines} lines — extract fetch/source modules"
+
+
 # CD4
 def test_quota_check_no_silent_broad_except():
     """CC-2: quota_guard.py must not have bare 'except Exception: sys.exit(0)'.
